Lectures/12Forrásállományok-20251202/requirements.txt
pycryptodome>=3.19.0

